Fix uart overrun issue.
The problem was that ALL interrupts are disabled when any .lock() is
issued inside any task. By rearranging priorites, we now don't need a
lock on terminal.render(). So the critical sections are minimized so
that the uart interrupt is not unecessary blocked.

This is an issue with RTIC on Cortex M0.